Make VCD dump optional

+++ opc6/README.csh
@@ -1,7 +1,7 @@
 #!/bin/tcsh -f
 # Remove non primary data files
 pushd tests
-rm -rf *~ *sim *trace *vcd *dump `ls -1 | egrep -v '(\.v$|\.csh|\.ucf|\.py|\.s$|spartan|xc95|opc6system|opc6copro|Make*)'`
+rm -rf *~ *sim *trace *vcd *dump `ls -1 | egrep -v '(\.v$|\.csh|\.ucf|\.py|\.s$|spartan|xc95|opc6system|opc6copro|Make*|stdin)'`
 
 if ( $#argv > 0 ) then
     if ( $argv[1] == "clean" ) exit
@@ -24,18 +24,21 @@ foreach test ( $testlist )
     echo "Running Test $test"
     # Assemble the test
     python3 ${assembler} ${test}.s ${test}.hex >  ${test}.lst
-    # Run the emulator
-    ${pyexec} ../opc6emu.py ${test}.hex ${test}.dump > ${test}.trace
+    
+    ${pyexec} ../opc6emu.py ${test}.hex ${test}.dump  > ${test}.trace
     # Test bench expects the hex file to be called 'test.hex'
     cp ${test}.hex test.hex
-    # Run icarus verilog to compile the testbench
-
+    # Run icarus verilog to compile the testbench only if there is no stdin file
     foreach option ( NEGEDGE_MEMORY POSEDGE_MEMORY )
         iverilog -D_simulation=1 -D${option}=1 ../opc6tb.v ../opc6cpu.v
+        # -D_dumpvcd=1
+        
         # Execute the test bench
         ./a.out | tee ${test}_${option}.sim
         # Save the results
-        mv dump.vcd ${test}_${option}.vcd
+        if ( -e dump.vcd) then
+            mv dump.vcd ${test}_${option}.vcd
+        endif            
         mv test.vdump ${test}_${option}.vdump
     end
 end
@@ -44,13 +47,15 @@ echo ""
 echo "Comparing memory dumps between emulation and simulation"
 echo "-------------------------------------------------------"
 foreach test ( $testlist )
-    foreach option ( NEGEDGE_MEMORY POSEDGE_MEMORY )
-        printf "%32s :" ${test}_${option}
-        if "${test}" =~ "*int" then
-            python3 ../../utils/mdumpcheck.py ${test}.dump  ${test}_${option}.vdump 0xF000 0x0500 0xFFFF
-        else
-            python3 ../../utils/mdumpcheck.py ${test}.dump  ${test}_${option}.vdump
-        endif
-    end
+    if ( ! -e $test.stdin ) then
+        foreach option ( NEGEDGE_MEMORY POSEDGE_MEMORY )
+            printf "%32s :" ${test}_${option}
+            if "${test}" =~ "*int" then
+                python3 ../../utils/mdumpcheck.py ${test}.dump  ${test}_${option}.vdump 0xF000 0x0500 0xFFFF
+            else
+                python3 ../../utils/mdumpcheck.py ${test}.dump  ${test}_${option}.vdump
+            endif
+        end
+    endif        
 end
 popd
